+++ src/components/Dragonfly/Dragonfly.js
@@ -5,7 +5,7 @@ import './dragonfly.css';
 /**
  * Import actions here
  */
-import {getSelectedNode} from '../../reducers/dragonflyReducer';
+import {getSelectedNode, getSinks} from '../../reducers/dragonflyReducer';
 
 const dragonfly = React.createClass({
     getInitialState: function(){
@@ -16,13 +16,14 @@ const dragonfly = React.createClass({
         return <div id="dragonfly">
         <div id="left"></div>
         <div id="center" key="2">{ this.props.selectedNode.hash }</div>
-        <div id="right"></div>
+        <div id="right"><ul>{ this.props.sinks.map(v => <li key={v.hash}>{v.hash}</li>)}</ul></div>
         </div>
     }
 });
 
 const mapStateToProps = state => ({
-    selectedNode: getSelectedNode(state)
+    selectedNode: getSelectedNode(state),
+    sinks: getSinks(state),
 });
 
 const mapDispatchToProps = dispatch => ({