+++ src/reducers/dragonflyReducer.js
@@ -1,6 +1,6 @@
 import {combineReducers} from 'redux';
 
-import {SELECT_NODE} from '../actions/dragonflyActions';
+import {SELECT_NODE, GET_SINKS} from '../actions/dragonflyActions';
 
 /**
  * Reducer for selecting 
@@ -14,12 +14,26 @@ const selectedNodeReducer = (node = {}, action) => {
     }
 }
 
+// Selector for getting the selected node.
 export const getSelectedNode = state => state.dragonfly.selectedNode;
 
+const sinksReducer = (sinks = [], action) => {
+    switch (action.type){
+        case GET_SINKS:
+            return [...action.nodes]
+        default:
+            return sinks
+    }
+}
 
+/**
+ * Selector for the dragonfly sinks.
+ */
+export const getSinks = state => state.dragonfly.sinks;
 
 const dragonflyReducer = combineReducers({
-    selectedNode: selectedNodeReducer
+    selectedNode: selectedNodeReducer,
+    sinks: sinksReducer
 })
 
 export default dragonflyReducer;